move shared code to letsencrypt mod
